backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Darshan <karkeradarshan98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
gunicorn = "^20.0"  # Ensure this line is present
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
pdfminer-six = "^20240706"
langchain = "^0.3.5"
openai = "0.28"
pinecone-client = "^5.0.1"
pymupdf = "^1.24.13"
python-multipart = "^0.0.17"
transformers = "^4.46.2"
torch = "^2.5.1"
sentencepiece = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"